controller and button separated
as well as some small changes (not working yet with midi, open port list not staying)
